Possibilité de voir les candidatures spontanées

+++ dashboard/asked-offers/asked-offers.php
@@ -1,18 +1,23 @@
 <h2 class="margin-top-none" style="display: inline-block">Candidatures</h2>
 
+<div class="" style="margin-bottom: 15px;">
+  <label for="prompt">Afficher les candidatures spontanées </label>
+  <input type="checkbox" name="prompt" id="prompt" value="" <?php if(isset($_GET["prompt"])){echo "checked";} ?>>
+</div>
+
 <?php
-$candidacies = $candidacy->getCandidacys();
+$candidacies = !isset($_GET["prompt"]) ? $candidacy->getCandidacys() :  $alert->getAlerts();
 ?>
 
 <span class="stat" style="display: block">
   <?php
   if(count($candidacies) > 1){
-    echo count($candidacies)." demandes trouvées";
+    echo count($candidacies)." candidatures trouvées";
   }else {
     if(count($candidacies) > 0){
-      echo "Une demande trouvée";
+      echo "Une candidature trouvée";
     }else {
-      echo "Aucun demande trouvée";
+      echo "Aucun candidature trouvée";
     }
   }
   ?>
@@ -29,11 +34,12 @@ $candidacies = $candidacy->getCandidacys();
         <span class="suggest_title"><i class="material-icons vertical-align-bottom margin-right-5 background-primary">person</i> <?php echo htmlspecialchars_decode($candidacy->getName())." ".htmlspecialchars_decode($candidacy->getFirst_name()); ?></span>
       </div>
       <div class="suggest_row">
-        <span class="suggest_col" style="max-width: 100%; overflow: hidden; height: 1.4em;" title="Domaines : <?php echo str_replace(",", " - ", htmlspecialchars_decode($candidacy->getDomains())); ?>">
+        <span class="suggest_col" style="max-width: 100%; overflow: hidden; height: 1.4em;" title="Domaines : <?php echo str_replace(",", " - ", htmlspecialchars_decode(isset($_GET["prompt"]) ? $candidacy->getDomain() : $candidacy->getDomains())); ?>">
           <i class="material-icons vertical-align-bottom margin-right-5 background-primary">domain</i>
-          Domaines : <?php echo str_replace(",", " - ", htmlspecialchars_decode($candidacy->getDomains())); ?>
+          Domaines : <?php echo str_replace(",", " - ", htmlspecialchars_decode(isset($_GET["prompt"]) ? $candidacy->getDomain() : $candidacy->getDomains())); ?>
         </span>
       </div>
+      <?php if(!isset($_GET["prompt"])){ ?>
       <div class="suggest_row">
         <span class="suggest_row">Intéréssé par l'offre No <?php echo $candidacy->getId_offer(); ?></span>
         <a href="<?php echo _ROOT_PATH."job/offers/a/Offres d'emploi/".$candidacy->getId_offer()."/" ?>" target="_blank" class="btnEdit" id="" title="Voir l'offre">
@@ -41,6 +47,7 @@ $candidacies = $candidacy->getCandidacys();
           <i class="material-icons vertical-align-bottom">launch</i>
         </a>
       </div>
+      <?php } ?>
       <div class="suggest_row">
         <?php if($candidacy->getCv_file() != ""){ ?>
         <span class="btnEdit" id="btnSeeCV<?php echo $candidacy->getId(); ?>" title="Voir le CV" style="margin-left: 5px">
@@ -48,12 +55,16 @@ $candidacies = $candidacy->getCandidacys();
           CV
         </span>
         <?php } ?>
-        <?php if($candidacy->getMotivation_file() != ""){ ?>
+        <?php
+          if(!isset($_GET["prompt"])){
+            if($candidacy->getMotivation_file() != ""){
+        ?>
         <span class="btnEdit" id="btnSeeMotivation<?php echo $candidacy->getId(); ?>" title="Voir la lettre de motivation">
           <i class="material-icons vertical-align-bottom">insert_drive_file</i>
           Lettre de motivation
         </span>
-        <?php } ?>
+        <?php }} ?>
+
         <span class="btnDelete" id="btnDeleteCandidacy<?php echo $candidacy->getId(); ?>" title="Supprimer cette candidature">
           <i class="material-icons vertical-align-bottom">close</i>
         </span>
@@ -70,8 +81,11 @@ $candidacies = $candidacy->getCandidacys();
       <div class="item_modal_body">
       <!--  <h3 style="font-weight: normal">Candidature <strong>No <?php echo $candidacy->getId(); ?></strong>, posté <?php get_elapsed_time($candidacy->getAdded_at()); ?></h3> -->
 
+      <?php if(!isset($_GET["prompt"])){ ?>
         <embed src="<?php echo _ROOT_PATH."files/candidacy/".$candidacy->getId()."/".htmlspecialchars_decode($candidacy->getCv_file()) ?>" width="100%" height="98%" type="application/pdf" />
-
+      <?php }else{ ?>
+        <embed src="<?php echo _ROOT_PATH."files/alerts/".$candidacy->getId()."/".htmlspecialchars_decode($candidacy->getCv_file()) ?>" width="100%" height="98%" type="application/pdf" />
+      <?php } ?>
       </div>
 
       <div class="item_modal_header">
@@ -83,6 +97,7 @@ $candidacies = $candidacy->getCandidacys();
   </div>
 
   <?php // See Motivation Modal ?>
+  <?php if(!isset($_GET["prompt"])){ ?>
   <div class="item_modal_shadow" id="seeMotivationModal<?php echo $candidacy->getId(); ?>">
     <div class="item_modal">
 
@@ -100,6 +115,7 @@ $candidacies = $candidacy->getCandidacys();
       </div>
     </div>
   </div>
+  <?php } ?>
 
   <input type="hidden" name="" value="<?php echo $candidacy->getId(); ?>"  id="id<?php echo $candidacy->getId(); ?>">
 
@@ -126,6 +142,14 @@ $candidacies = $candidacy->getCandidacys();
 
 <script type="text/javascript">
   $(document).ready(function(){
+    $("#prompt").change(function(){
+      if($("#prompt").prop("checked")){
+        window.location.assign(_ROOT_PATH+"dashboard/asked-offers/p/prompt");
+      }else {
+        window.location.assign(_ROOT_PATH+"dashboard/asked-offers/");
+      }
+    });
+
   	//Show or hide modal
    function toggleModal(modal, modal_btn, modal_btn_close){
      $(".item_deleteModal, .item_modal").click(function(ev){
@@ -163,7 +187,11 @@ $candidacies = $candidacy->getCandidacys();
       //for delete
       toggleModal($("#deleteCandidacyModal<?php echo $candidacy->getId(); ?>"), $("#btnDeleteCandidacy<?php echo $candidacy->getId(); ?>"), $("#btnDeleteCandidacyClose<?php echo $candidacy->getId(); ?>, .item_deleteModal_shadow"));
 
-      deleteCandidacyInBD($("#btnDeleteCandidacyConfirm<?php echo $candidacy->getId(); ?>"), "outils/php/traitement/candidacy/delete-candidacy.php", $("#id<?php echo $candidacy->getId(); ?>"));
+      <?php if(!isset($_GET["prompt"])){ ?>
+        deleteCandidacyInBD($("#btnDeleteCandidacyConfirm<?php echo $candidacy->getId(); ?>"), "outils/php/traitement/candidacy/delete-candidacy.php", $("#id<?php echo $candidacy->getId(); ?>"));
+      <?php }else{ ?>
+        deleteCandidacyInBD($("#btnDeleteCandidacyConfirm<?php echo $candidacy->getId(); ?>"), "outils/php/traitement/candidacy/delete-alert.php", $("#id<?php echo $candidacy->getId(); ?>"));
+      <?php } ?>
 
       //for edit
       toggleModal($("#seeCVModal<?php echo $candidacy->getId(); ?>"), $("#btnSeeCV<?php echo $candidacy->getId(); ?>"), $("#btnSeeCVClose<?php echo $candidacy->getId(); ?>, .item_modal_shadow"));
